adds search function to browseActivity
